requirements.txt
astroid==2.4.2
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==2.2.14
django-extensions==3.1.0
django-heroku==0.3.1
django-rest-swagger==2.2.0
django-slugify-processor==0.9.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.3.0
flake8==3.8.4
gunicorn==20.0.4
idna==2.10
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
openapi-codec==1.3.2
Pillow==8.1.0
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
pylint==2.6.0
python-dateutil==2.4.2
python-decouple==3.4
pytz==2020.4
requests==2.24.0
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
tldextract==1.6
toml==0.10.2
unicodecsv==0.13.0
uritemplate==3.0.1
urllib3==1.25.11
whitenoise==5.2.0
wrapt==1.12.1
